pretty much done
Bye bye maybe forever? IDK if I'll ever use this github account again. Maybe. IDK if I'll ever finish this project or not. Maybe chopped. Life happens. Things come and go. This is one of them. I mean. This was something I did every day for 4 years? That is nearly a whole THIRD of my life. All that on this one thing in my life. Farewell.